Split up test file between suites
